Layer5 svg added + width to 100% w/signoff
